Make MyMoneyFile header less dependent
The goal of this patch is to reduce dependability on mymoneyfile and of
mymoneyfile on other libraries.
Changes:
1) no header file includes mymoneyfile now,
2) mymoneyfile.h uses forward-declarations instead of includes.

Patch is huge but it's nothing but code shuffling. Nonetheless it has
two important merits i.e. smaller executable file and shorter
compilation times.

Differential Revision: https://phabricator.kde.org/D8535

+++ kmymoney/mymoney/storage/mymoneydbdef.cpp
@@ -540,7 +540,7 @@ const QString MyMoneyDbDef::generateSQL(const QExplicitlySharedDataPointer<MyMon
       toReplace.prepend(':');
       QString replace = "NULL";
       if ((*act)->name() == "accountType")
-        replace = QString::number(pac->accountType());
+        replace = QString::number((int)pac->accountType());
       if ((*act)->name() == "accountTypeString")
         replace = enclose(pac->name());
       if ((*act)->name() == "isStockAccount")
